+++ ConsoleApp1/TradeResult.cs
@@ -8,8 +8,8 @@ namespace ConsoleApp1
     {
         //properties (get and set) - used a lot for holding values
 
-        public int TotalBitcoin { get; set; }
-        public int TotalMoney { get; set; }
+        public float TotalBitcoin { get; set; }
+        public float TotalMoney { get; set; }
 
 
         //data
